backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.0
beanie==1.23.6
dnspython==2.4.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
pandas==2.1.3
numpy==1.26.2
statsmodels==0.14.1
joblib==1.3.2
shap==0.44.0

# Data Processing
geopy==2.4.1
scipy==1.11.4

# External APIs
httpx==0.25.2
aiohttp==3.9.1

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
fpdf2==2.7.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging
loguru==0.7.2
